Deleted LightweightUtils. Moved work to Core (partial class). Normalized files with the autoformat tool to be consistent to the convention.
