--- ScriptableRenderPipeline/LightweightPipeline/LWRP/LightweightPipelineUtils.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 1556a6d19597541c8bed4b90c704fb06
-timeCreated: 1505115136
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
