+++ ScriptableRenderPipeline/LightweightPipeline/LWRP/SceneViewDrawMode.cs
@@ -1,4 +1,4 @@
-﻿#if UNITY_EDITOR
+#if UNITY_EDITOR
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -24,7 +24,7 @@ namespace UnityEditor.Experimental.Rendering.LightweightPipeline
                 cameraMode.drawMode == DrawCameraMode.ValidateMetalSpecular ||
                 cameraMode.drawMode == DrawCameraMode.ShadowMasks ||
                 cameraMode.drawMode == DrawCameraMode.LightOverlap
-            )
+                )
                 return false;
 
             return true;
